refactor: split config into package
